requirements.txt
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cycler==0.10.0
dataclasses==0.8
Flask==2.0.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mpld3==0.5.5
numpy==1.19.5
packaging==21.0
pandas==1.1.5
Pillow==8.3.2
pkg-resources==0.0.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
waitress==2.0.0
Werkzeug==2.0.2
zipp==3.6.0
